Move rust flip-flop function into its own module
Keep a python wrapper of the function in lib.rs
