+++ pyproject.toml
@@ -66,7 +66,7 @@ dev =[
 profile = "release"
 features = ["pyo3/extension-module"]
 manifest-path = "rust/Cargo.toml"
-module-name = "lipyphilic._lipyferrous"
+module-name = "lipyphilic._lipyferric"
 python-source = "src"
 python-packages = ["lipyphilic"]
 strip = true
