+++ rust/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
-name = "_lipyferrous"
+name = "_lipyferric"
 path = "src/lib.rs"
 crate-type = ["cdylib", "rlib"]
